API: rename pandas/core/generic.py -> pandas/core.ndframe.py
