+++ pandas/io/formats/style.py
@@ -48,7 +48,7 @@ from pandas.core.frame import (
     DataFrame,
     Series,
 )
-from pandas.core.generic import NDFrame
+from pandas.core.ndframe import NDFrame
 from pandas.core.shared_docs import _shared_docs
 
 from pandas.io.formats.format import save_to_buffer
